Problem: caching certain calls within statement
There are cases where it is important to get stable values for certain
calls (like repeated API calls) for consistency and latency reasons.

Imagine we call an API for a view and then join with another view, that
makes that very same request.

Solution: introduce statement/portal-level scoped variables

These go away once portal has changed.
