+++ extensions/omni_var/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - TBA
+## [0.3.0] - TBA
+
+### Added
+
+* Statement/portal-scoped variables [#680](https://github.com/omnigres/omnigres/pull/680)
 
 ## [0.2.2] - 2024-10-22
 
@@ -38,3 +42,5 @@ Initial release following a few months of iterative development.
 [0.2.1]: [https://github.com/omnigres/omnigres/pull/619]
 
 [0.2.2]: [https://github.com/omnigres/omnigres/pull/652]
+
+[0.3.0]: [https://github.com/omnigres/omnigres/pull/679]
