+++ versions.txt
@@ -20,7 +20,7 @@ omni_session=0.1.0
 omni_sql=0.3.3
 omni_txn=0.4.0
 omni_types=0.2.0
-omni_var=0.2.2
+omni_var=0.3.0
 omni_vfs_types_v1=0.1.0
 omni_vfs=0.1.2
 omni_web=0.1.1
